refac(virtual): DRY language service ceration
