+++ server/src/virtual/virtualDefinition/virtualDefinition.ts
@@ -4,45 +4,10 @@ import { DefinitionResult } from "../../feature/definition/getDefinition";
 import { AureliaProgram } from "../../viewModel/AureliaProgram";
 import {
   createVirtualViewModelSourceFile,
+  getVirtualLangagueService,
   VirtualSourceFileInfo,
   VIRTUAL_SOURCE_FILENAME,
 } from "../virtualSourceFile";
-
-export function getVirtualLangagueService(
-  sourceFile: ts.SourceFile
-): ts.LanguageService {
-  let compilerSettings = {} as ts.CompilerOptions;
-  compilerSettings = {
-    module: ts.ModuleKind.CommonJS,
-    target: ts.ScriptTarget.ESNext,
-    outDir: "dist",
-    emitDecoratorMetadata: true,
-    experimentalDecorators: true,
-    lib: ["es2017.object", "es7", "dom"],
-    sourceMap: true,
-    rootDir: ".",
-  };
-
-  const host: ts.LanguageServiceHost = {
-    getCompilationSettings: () => compilerSettings,
-    getScriptFileNames: () => [sourceFile.fileName],
-    getScriptVersion: () => "0",
-    getScriptSnapshot: () => ts.ScriptSnapshot.fromString(sourceFile.getText()),
-    getCurrentDirectory: () => process.cwd(),
-    getDefaultLibFileName: (options) => ts.getDefaultLibFilePath(options),
-    fileExists: ts.sys.fileExists,
-    readFile: ts.sys.readFile,
-    readDirectory: ts.sys.readDirectory,
-  };
-  const cls = ts.createLanguageService(host, ts.createDocumentRegistry());
-
-  if (!cls) {
-    console.log("No cls");
-    throw new Error("no cles");
-  }
-
-  return cls;
-}
 
 export function createVirtualFileWithContent(
   aureliaProgram: AureliaProgram,
